Entity 속성 타입 Long -> Integer 변경 - 민호

+++ src/main/java/hot/member/domain/Estimate.java
@@ -41,7 +41,7 @@ public class Estimate {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "est_no")
-	private Long estNo;
+	private Integer estNo;
 	
 	@ManyToOne
 	@JoinColumn(name="member_no")
